+++ src/plugin/css.js
@@ -38,17 +38,14 @@
 /*if ($P.WORKER || $P.NODE) {*///#JSCOVERAGE_IF 0/*}*/
         // Create stylesheet link
         var link = sb.document.createElement("link"),
-            id = +new sb.global.Date,
-            onload = function (e) {
+            id = Math.random() + '',
+            onload = function () {
                 if (isNotLoaded) {
                     isNotLoaded = 0;
                     // register or cleanup
                     link.removeAttribute('id');
 
-                    if (!e) {
-                        sb.trigger('*:request-error', moduleName, module);
-                    }
-                    callback(e ? sb.register(moduleName, link) : head.removeChild(link) && sb.undefined); // e === undefined if error
+                    callback(sb.register(moduleName, link)); // e === undefined if error
                 }
             };
 
@@ -57,10 +54,21 @@
         link.rel = "stylesheet";
         link.id = id;
 
-        sb.global.setTimeout(onload, 3000, 0);
-
         head = sb.document.getElementsByTagName("head")[0];
         head.insertBefore(link, head.firstChild);
+
+        function isRules(sheet) {
+            if ((sheet.ownerNode || sheet.owningElement).id != id) {
+                return false;
+            }
+            try {
+                // It can be null or throw an Security error in case of cross origin stylesheets
+                return !!(sheet.cssRules || sheet.rules).length;
+            } catch (e) {
+                // In case of access error assume that css is loaded
+                return true;
+            }
+        }
 
         (function poll() {
             if (isNotLoaded) {
@@ -70,17 +78,15 @@
 
                 try {
                     for (; j < k; j++) {
-                        if((sheets[j].ownerNode || sheets[j].owningElement).id == id &&
-                            (sheets[j].cssRules || sheets[j].rules).length) {
+                        if (isRules(sheets[j])) {
 //#JSCOVERAGE_IF 0
-                            return onload(1);
+                            return onload();
 //#JSCOVERAGE_ENDIF
                         }
                     }
                     // if we get here, its not in document.styleSheets (we never saw the ID)
                     throw 1;
                 } catch(e) {
-                    /*if ($P.FILE_PROTOCOL) {*/if (e != 1 && (e.code === 15 || sheets[j].cssRules === null)) {return onload(1);}/*}*/
                     // Keep polling
                     sb.global.setTimeout(poll, 90);
                 }